TRE-971 Tech Debt (#16)
* TRE-971 Deleted unused ISODateFormatter and ISODateFormatterSpec

* TRE-971 clearing unused imports and variables

* TRE-971 changing availableReportsCount to Int

* TRE-971 Refactor getReportType avoiding contains in string comparison

* TRE-971 log missing headers and add loggers in EoriUpdateController

* TRE-971 use isSuccessful and isServerError methods of Status

* TRE-971 scala formatting

* TRE-971 replace hardcoded eori string with applicationconstants value

+++ app/uk/gov/hmrc/tradereportingextracts/repositories/ReportRequestRepository.scala
@@ -130,7 +130,7 @@ class ReportRequestRepository @Inject() (appConfig: AppConfig, mongoComponent: M
         .map(_.filter(_.isReportStatusComplete))
     }
 
-  def countAvailableReports(eori: String)(using ec: ExecutionContext): Future[Long] = Mdc.preservingMdc {
+  def countAvailableReports(eori: String)(using ec: ExecutionContext): Future[Int] = Mdc.preservingMdc {
     collection
       .find(Filters.equal("requesterEORI", eori))
       .toFuture()
